Finished refactoring symbol parsing
