+++ src/dawn/analysis/Scope.ts
